Move drawable camera to generic components folder
