+++ frontend/src/components/ui/DrawableCamera/DrawableCamera.tsx
