+++ frontend/src/components/ui/DrawableCamera/index.ts
@@ -1,0 +1,1 @@
+export * from "./DrawableCamera";
